# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ResponseStatus": {
				"enum": [
					"success",
					"error"
				],
				"type": "string"
			},
			"ResponseDTO_unknown_": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/ResponseStatus"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserProfile": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"countryCode": {
						"type": "string"
					},
					"profile": {
						"properties": {
							"emailVerified": {
								"type": "boolean"
							},
							"phoneVerified": {
								"type": "boolean"
							},
							"lastLogin": {
								"type": "string",
								"format": "date-time",
								"nullable": true
							},
							"address": {
								"type": "string",
								"nullable": true
							},
							"phoneNumber": {
								"type": "string",
								"nullable": true
							},
							"nickname": {
								"type": "string",
								"nullable": true
							},
							"role": {
								"type": "string"
							}
						},
						"required": [
							"lastLogin",
							"address",
							"phoneNumber",
							"nickname",
							"role"
						],
						"type": "object",
						"nullable": true
					},
					"paymentMethod": {
						"items": {
							"properties": {
								"institution": {
									"type": "string"
								},
								"paymentId": {
									"type": "string"
								},
								"name": {
									"type": "string"
								},
								"id": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"institution",
								"paymentId",
								"name",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					},
					"transactionsCount": {
						"type": "number",
						"format": "double"
					},
					"fraudScores": {
						"items": {
							"properties": {
								"result": {
									"type": "string"
								},
								"score": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"result",
								"score"
							],
							"type": "object"
						},
						"type": "array"
					},
					"apiKey": {
						"properties": {
							"lastUsed": {
								"type": "string",
								"format": "date-time",
								"nullable": true
							},
							"createdAt": {
								"type": "string",
								"format": "date-time"
							},
							"key": {
								"type": "string"
							}
						},
						"required": [
							"lastUsed",
							"createdAt",
							"key"
						],
						"type": "object",
						"nullable": true
					}
				},
				"required": [
					"id",
					"firstName",
					"lastName",
					"email",
					"type",
					"status",
					"countryCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseDTO_IUserProfile_": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/ResponseStatus"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/IUserProfile"
					}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseDTO_IUserProfile-Array_": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/ResponseStatus"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/IUserProfile"
						},
						"type": "array"
					}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseDTO__Name-string--Email-string__": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/ResponseStatus"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"Email": {
								"type": "string"
							},
							"Name": {
								"type": "string"
							}
						},
						"required": [
							"Email",
							"Name"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateUserDTO": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"countryCode": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email",
					"password",
					"type",
					"status",
					"countryCode"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "asgard",
		"version": "1.0.0",
		"description": "Vijipay's Webservice",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
		"/ag/v1/user/{userId}/disable": {
			"patch": {
				"operationId": "DisableUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/ag/v1/user/find": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseDTO_IUserProfile_"
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "email",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/ag/v1/user/find-by-id": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseDTO_IUserProfile_"
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/ag/v1/user/find-by-email": {
			"get": {
				"operationId": "GetUserByEmail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseDTO_IUserProfile_"
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/ag/v1/user/all": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseDTO_IUserProfile-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [],
				"parameters": []
			}
		},
		"/ag/v1/user/{userId}/delete": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/ag/v1/user/deleteAll": {
			"delete": {
				"operationId": "DeleteAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [],
				"parameters": []
			}
		},
		"/ag/v1/users/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseDTO__Name-string--Email-string__"
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDTO"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}